FIX : Place lat,lng 타입 변경으로 인한 수정

+++ matching/src/main/java/com/ssafy/matching/dto/Place.java
@@ -23,7 +23,7 @@ public class Place implements Serializable {
     private int placeId;
     private String address;
     @ApiModelProperty(value = "위도")
-    private Float lat;
+    private double lat;
     @ApiModelProperty(value = "경도")
-    private Float lng;
+    private double lng;
 }
